Backend/requirements.txt
langgraph
langchain_groq
langchain_huggingface
langchain
langchain-community
pydantic
requests
python-dotenv
fastapi
uvicorn
httpx
psycopg2
sentence-transformers
faiss-cpu
pypdf
python-socketio[client]
elevenlabs
gtts
pydub
python-multipart
twilio
twilio[voice]
websockets
requests